Added more threads to car audio HAL.
Audio HAL did not have enough threads for all the work available. Each
input and ouput devices starts a thread for interacting with the PCM
and a binder thread for comunicating. Thus the pools of thread was
running out. Added more threads to the AIDL side for audio control.

Also added names to the input devices. Set the name of the thread
according to the input/output device name for better identification.

Removed the un-needed .rc command to restart the audioserver in case of
audio HAL crash. The audioserver already has mechanism to re-start
itself in case of failure. It will also attempt to re-start the audio
HAL if the HAL does crashes.

Modified how mixer thread is terminated, it used to kill the thread.
Now the loop is attempted to stop and join from the close PCM method, if
that fails the thread is still killed.

Fixed the format of the audio policy configuration xml file so that all
lines longer than 100 characters are aligned to previous's line first
attribute.

Bug: 215747045
Test: m -j, and run emulator
Test: atest VtsAidlHalAudioControlTest
Change-Id: I18a077719cf6b0d6db2ef3d3c723bdc325da4ac4